Added new styles and updated home page to fit the forum description.
I used an online resource to generate the color palette for the website, and I think that it looks super good along with the css I added for the home page. The nav bar is still a little empty, but I hope to be able to update that in the future.